Add implementation to mime fullversion support
We just introduced fullversion support in CodiMD[1][2]. This allows us
to inform all clients about every update of a container.

This patch adds some container specific tricks to keep the current
commit and repository available for this fullversion script, so whenever
someone uses those Dockerfiles to build a container, the source URL for
the CodiMD from this container is correct.

[1]: https://github.com/hackmdio/codimd/pull/991
[2]: https://github.com/hackmdio/codimd/commit/bcc914a7735290b86e0df428aed75dabdf9f1eca