Add script to merge master before testing
Travis CI provides feedback about a change at the timeit's purpose and
how it works in it's own system. What it doesn't tell is if it's
breaking our master branch or not.

This patch adds some bash scripting that merges master before running
tests on PR to make sure it doesn't break the master branch. This will
also allow to fix the master and then merge other changes without the
need to rebase the changes to validate it works now.