feat: add the ability to choose configuration type
Add the ability to choose multiple type of configuration. It could be a
configuration using CMD_DB_URL (which was the standard way up to now).

But it is now possible to send multiple environment variable to
configure the database.

This makes the container easier to configure on many platform (including
k8s) as the managed database is sometime offering credential as indivual
variable instead of a long url.

Signed-off-by: oupala <oupala@users.noreply.github.com>