Fix tests by introducing a delay
This patch should fix current failing builds by adding a 30 seconds
delay between the container starting and the tests starting to run, so
that database migrations and alike can take place.

Currently the tests are running directly after the container was
started, which results in failing HTTP request. This can be explained by
a not-yet-ready hedgedoc instance as it's probably still doing the
database migrations and getting up for the first start.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>