Remove dockerize and entrypoint-script parsing
Currently the entrypoint script parses the database URL configuration
string in order to extract the remote location and waits for the
database to be available.

This patch removes this logic and also removes `dockerize`, which was
used to wait for the connection. However, this change should not be
merged without the changes introduced in a commit on the main
applicaiton, that implements an own connection retry mechanism.

Reference:
https://github.com/hedgedoc/hedgedoc/pull/1228

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>