Implement automatic version management for releases
This patch enables renovate to notice new hedgedoc releases and
automatically provide a PR that updates the hedgedoc release version.

This was implemented using the regex-manager that allows to parse
strings and extract the required values for datasource, versioning
schema and project name in order to have them handled by renovatebot.

Reference:
https://docs.renovatebot.com/modules/manager/regex/

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>