ci(github-actions): Build multi-arch images for nightly builds
This patch enabled multi-arch builds for nightly builds, this should
help to deploy hedgedoc on devices like a Raspberry Pi without resorting
to third-party images.

Initially, this should be done by nightly images, to evaluate if this
works properly. It can later-on be enabled on the general releases as
well.

Signed-off-by: Sheogorath <sheogorath@shivering-isles.com>